tui: Move session quit handling
Interactive mode now delegates help display, flow parsing, command escapes,
scoped menus, and hunk routing outside the main dispatcher.

Smart quit handling still reads session baselines, checks HEAD and index
state, prompts users, and stops or aborts sessions inside interactive.py.

This commit addresses that by moving smart quit handling into
tui.session_quit and importing handle_quit for the q action.

Subsequent commits will retarget quit tests and guard the session quit
owner.
